Address all linting errors
