+++ www/app/auth/page.tsx
@@ -26,7 +26,7 @@ export default function Auth() {
         router.push("/auth/reset")
       }
     })
-  }, [])
+  }, [router, supabase])
 
   return (
     <section className="bg-white">
@@ -34,10 +34,11 @@ export default function Auth() {
         <aside
           className="relative block h-16 lg:order-last lg:col-span-5 lg:h-full xl:col-span-6"
         >
-          <img
+          <Image
             alt="Pattern"
             src="/auth_banner.jpg"
             className="absolute inset-0 h-full w-full object-cover"
+            fill={true}
           />
         </aside>
