+++ www/package.json
@@ -32,6 +32,7 @@
   },
   "devDependencies": {
     "@types/react-syntax-highlighter": "^15.5.7",
+    "encoding": "^0.1.13",
     "eslint": "^8.49.0",
     "eslint-config-next": "^13.4.19"
   }
